Prevent fetching all RSS feeds while reading only 1 article
